Fixed the MiniGo Tests due to recent changes.

+++ Tests/MiniGoTests/Models/GoModelTests.swift
@@ -11,7 +11,7 @@ final class GoModelTests: XCTestCase {
         let sampleInput = Tensor<Float>(randomUniform: [2, 19, 19, 17])
         let inference = model.prediction(input: sampleInput)
         let (policy, value) = (inference.policy, inference.value)
-        XCTAssertEqual(TensorShape([2, Int32(19 * 19 + 1)]), policy.shape)
+        XCTAssertEqual(TensorShape([2, 19 * 19 + 1]), policy.shape)
         XCTAssertEqual(TensorShape([2]), value.shape)
     }
 }
